Commit Details
Correction des pages details pour Client et Vehicle

+++ src/main/java/com/epf/rentmanager/service/ReservationService.java
@@ -54,7 +54,7 @@ public class ReservationService {
 		return null;
 	}
 	
-	public List<Reservation> findReservationByVehiculeId(long id) throws ServiceException {
+	public List<Reservation> findReservationByVehiculeId(int id) throws ServiceException {
 		
 		try {
 			return this.reservationDao.findResaByVehicleId(id);
